chore: rocks: refactor types.rs into types module
