+++ src/eth/storage/rocks/types/transaction_mined.rs
@@ -1,0 +1,48 @@
+use std::fmt::Debug;
+
+
+use crate::eth::primitives::EvmExecution;
+use crate::eth::primitives::LogMined;
+use crate::eth::primitives::TransactionMined;
+
+use super::block_number::BlockNumberRocksdb;
+use super::hash::HashRocksdb;
+use super::index::IndexRocksdb;
+use super::log_mined::LogMinedRockdb;
+use super::transaction_input::TransactionInputRocksdb;
+
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct TransactionMinedRocksdb {
+    pub input: TransactionInputRocksdb,
+    pub execution: EvmExecution,
+    pub logs: Vec<LogMinedRockdb>,
+    pub transaction_index: IndexRocksdb,
+    pub block_number: BlockNumberRocksdb,
+    pub block_hash: HashRocksdb,
+}
+
+impl From<TransactionMined> for TransactionMinedRocksdb {
+    fn from(item: TransactionMined) -> Self {
+        Self {
+            input: item.input.into(),
+            execution: item.execution,
+            logs: item.logs.into_iter().map(LogMinedRockdb::from).collect(),
+            transaction_index: IndexRocksdb::from(item.transaction_index),
+            block_number: BlockNumberRocksdb::from(item.block_number),
+            block_hash: HashRocksdb::from(item.block_hash),
+        }
+    }
+}
+
+impl From<TransactionMinedRocksdb> for TransactionMined {
+    fn from(item: TransactionMinedRocksdb) -> Self {
+        Self {
+            input: item.input.into(),
+            execution: item.execution,
+            logs: item.logs.into_iter().map(LogMined::from).collect(),
+            transaction_index: item.transaction_index.into(),
+            block_number: item.block_number.into(),
+            block_hash: item.block_hash.into(),
+        }
+    }
+}
